wip, rename PublicAddress to EthAddress
