--- example/src/stories/PublicAddress/docs.md
@@ -1,11 +1,0 @@
-# PublicAddress
-
-**Import component:**
-
-```jsx
-import { PublicAddress } from 'rimble-ui';
-```
-
-**Output:**
-
-<!-- STORY -->
