+++ src/EthAddress/CopyToClipboard.js
